Added Void wrapping class and DoubleByte
Void Wrapper class serves the purpose of wrapping the void. When called as a type parameter, you can get the original void value from the VoidWrapper value. DoubleByte serves the purpose of creating a new datatype called "DoubleByte". The maximal value of DoubleByte is 255 and minimal -256. You can contain values in it by typing new DoubleByte(int i) and then complete it by adding .integerValue() to get the basic integer value from it.